made move function a coroutine to make it move in a tuk tuk manner
the moving backwards single bug remains pointed out in the comments of previous commit. i can pinpoint the why its happening in the move function now. but i was thinking of just recreating the move function all together
